Added Prescription request
This was by far the hardest and most complicated request to add. While the concept was simple enough, the changes required to implement the idea cascaded through classes. I wanted the PrescriptionRequest to almost work independently of a Secretary. This required it to implement the Observer pattern on the Drug's stock level (essentially). With the previous implementation of the DrugRepository, this was not easily done;  making the repository observable would have resulted in, after a stock update to a specific drug,  iterating over all observers, regardless of whether they want to know the stock of that drug or not. With the improved system, each prescription request is subscribed to each specific drug stock listing. This means that when the drug's stock is updated, it will only iterate through the prescriptions regarding that drug.

With this current implementation, it also lays the ground work to fully automate the prescription approval process as the Request's approve function could be called within the Observer's update function.

Finally, I updated the DrugRepositoryTest to reflect the changes in DrugRepository and ensure that is all still works as it should