Constructor not adding item to repository
A nuance in Java is that a constructor cannot reference an object is has not yet fully created (or I think that's how it works as it didn't work for me). Therefore, I introduced the I_RepositoryItem interface that shows an item belongs to a repository. If you want an object to be included in a repository, simple follow the object initialisation with the .include() function.