PrescriptionRequestTest
Being the most complicated test, I thought I should test it. The other tests are quite straight forward however. A couple of errors were changed, including the fact that the Observer pattern wasn't fully implemented, and when the Request repository was clearing, it was wiping the internal structure of the EnumMap.